fix CMake

+++ examples/simple_benchmark.cpp
@@ -68,7 +68,7 @@ int main(int argc, char * argv[])
     auto end = std::chrono::high_resolution_clock::now();
     auto duration = std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
 
-    printf("Workload Time %d us\n", duration);
+    printf("Workload Time %d us\n", (int)duration);
 
     if (db.status == templatedb::OPEN)
         db.close();
